8th bit was too much, there are only 7 modes.
Otherwise the bit for the “stay on” mode is ignored when sending and all modes are deactivated. As a result, when receiving in the `transformModeFromHex`, 0000000 is turned into 1111111, which reactivates all modes.
This fix ensures that when all modes are deactivated, only the “stay on” mode is reactivated (as it was probably intended).
The check in `transformModeFromHex` for 0000000 could even be removed, as the device no longer responds with 0.